Use FFmpeg for MVE decoding
It turns out that FFmpeg is actually able to decode the Interplay MVE
movies so we can drop the old MVE code and just use FFMpeg and the new
video player interface for that.

+++ code/cutscene/Decoder.h
@@ -61,7 +61,7 @@ struct AudioFrame {
 
 	int channels;
 
-	long rate;
+	int rate;
 };
 typedef FramePtr<AudioFrame> AudioFramePtr;
